# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NHM Product Dat Xe Backend API",
        "description": "API documentation for NHM Product Dat Xe Backend",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/vouchers": {
            "get": {
                "tags": [
                    "Finance"
                ],
                "summary": "Xem danh sách voucher (UC-21)",
                "description": "Lấy danh sách các voucher đang hoạt động, có thể lọc theo loại dịch vụ (ride/food).",
                "operationId": "d9cad419582a2a5e9d4e6b9db5dc1db7",
                "parameters": [
                    {
                        "name": "service_type",
                        "in": "query",
                        "description": "Lọc voucher theo dịch vụ. \"ride\": Chuyến xe, \"food\": Đồ ăn. Nếu để trống sẽ lấy tất cả.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "ride",
                                "food"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách voucher",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/VoucherResponse"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/vouchers/{id}": {
            "get": {
                "tags": [
                    "Finance"
                ],
                "summary": "Xem chi tiết voucher (UC-21)",
                "description": "Xem chi tiết điều kiện áp dụng và mô tả của một voucher cụ thể.",
                "operationId": "b3d3172229e26fef298c0c2e34bea47f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của voucher",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chi tiết voucher"
                    },
                    "404": {
                        "description": "Không tìm thấy voucher"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/vouchers/{id}/save": {
            "post": {
                "tags": [
                    "Finance"
                ],
                "summary": "Lưu voucher (UC-21 A5)",
                "description": "Lưu voucher vào ví cá nhân để sử dụng sau.",
                "operationId": "2bea8938b937a3f014910292a27c6173",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của voucher",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Đã lưu voucher thành công"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/vouchers/{id}/apply-quick": {
            "post": {
                "tags": [
                    "Finance"
                ],
                "summary": "Áp dụng nhanh voucher (UC-22)",
                "description": "Kiểm tra voucher và trả về màn hình đích tương ứng (Ride/Food/Delivery) để chuyển hướng.",
                "operationId": "5411b6dd9444e747702d69ca6fdbba40",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của voucher",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thông tin chuyển hướng và voucher"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "VoucherResponse": {
                "description": "Thông tin chi tiết mã giảm giá",
                "required": [
                    "id",
                    "code",
                    "service_type",
                    "discount_type",
                    "discount_value"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "1"
                    },
                    "code": {
                        "type": "string",
                        "example": "DEMO10"
                    },
                    "service_type": {
                        "description": "Loại dịch vụ áp dụng. 1: Chuyến xe, 2: Đồ ăn, 3: Đa dịch vụ, 4: Giao hàng, 5: Tất cả",
                        "type": "integer",
                        "example": 1
                    },
                    "discount_type": {
                        "description": "Loại giảm giá. 1: Giảm tiền mặt cố định, 2: Giảm theo phần trăm",
                        "type": "integer",
                        "example": 1
                    },
                    "discount_value": {
                        "type": "number",
                        "format": "float",
                        "example": 10000
                    },
                    "min_order_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 50000
                    },
                    "max_discount_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 50000,
                        "nullable": true
                    },
                    "valid_until": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "is_saved": {
                        "description": "Đã lưu vào ví hay chưa",
                        "type": "boolean",
                        "example": false
                    },
                    "status": {
                        "description": "Trạng thái: Available, Expired, Used",
                        "type": "string",
                        "example": "Available"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}